+++ database/migrations/2020_12_03_135951_create_payment_gold_table.php
@@ -24,7 +24,7 @@ class CreatePaymentGoldTable extends Migration
             $table->bigInteger('agent_id')->unsigned();
             $table ->foreign('agent_id')->references('id')->on('users');
 
-            $table->integer('gold_received')->nullable(false);
+            $table->double('gold_received')->nullable(false);
 
             $table->date('received_date')->nullable(false);
